# Improve newChunk performance